Fixed install process. CLI Application will now be able to recognize arguments

+++ bin/SecureSignIn-4.1a-linux-install/install.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
 
-rm /opt/Cryogen/SecureSignIn-*
-mkdir /opt/Cryogen
-rm /usr/bin/ssi
-cp SecureSignIn-4.1a-linux /opt/Cryogen/
-touch /usr/bin/ssi
-echo "#!/bin/bash" > /usr/bin/ssi
-echo "/opt/Cryogen/SecureSignIn-4.1a-linux" > /usr/bin/ssi
+if [ -f /usr/bin/ssi ]; then
+	rm /usr/bin/ssi
+	echo "A version of Secure Sign In detected and removed."
+fi
+cp SecureSignIn-4.1a-linux /usr/bin/ssi
 chmod +x /usr/bin/ssi
+echo "Secure Sign In v4.1.1a successfully installed."
